Customizing To My Environment
Instead of having such a generic starting point that should be
accommodating to most environments, I now have it set to my environment.
In other words, I added my Elixir specific tooling and configs, because
that is what I work in.

Hopefully it is still simple enough to understand, such that anyone can
remove unwanted plugins, and update configs to match one desired.